requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
Pillow
python-telegram-bot
